api/requirements.txt
astroid==2.11.2
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
dill==0.3.4
Flask==2.1.1
Flask-Cors==3.0.10
idna==3.3
importlib-metadata==4.11.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.1
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.8.0
pylint==2.13.5
pymongo==3.11.3
python-dotenv==0.20.0
requests==2.27.1
six==1.16.0
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0
